# sistema/backend/src/main/resources/swagger.yaml
openapi: 3.0.1
info:
  title: Corrige Online
  version: 1.0.0
paths:
...
  /alunos/{alunoId}/doacoes:
    get:
      tags:
        - Doacao
      operationId: getAlunoDonations
      summary: Extrato de doações recebidas por um aluno
      parameters:
        - name: alunoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Extrato econtrado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './dto/response/doacao-reponse.yaml#/components/schemas/DoacaoResponseDTO'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: './dto/response/error-response.yaml#/components/schemas/ErrorResponse'
  /professores/{professorId}/doacoes:
    get:
      tags:
        - Doacao
      operationId: getProfessorDonations
      summary: Extrato de doações feitas por um professor
      parameters:
        - name: professorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Extrato econtrado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './dto/response/doacao-reponse.yaml#/components/schemas/DoacaoResponseDTO'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: './dto/response/error-response.yaml#/components/schemas/ErrorResponse'
    post:
      tags:
        - Doacao
      operationId: donate
      summary: Fazer uma doação para um aluno
      parameters:
        - name: professorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './dto/request/doacao-request.yaml#/components/schemas/DoacaoRequestDTO'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: './dto/response/doacao-reponse.yaml#/components/schemas/DoacaoResponseDTO'
        '406':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: './dto/response/error-response.yaml#/components/schemas/ErrorResponse'